Move image trust related cli methods into the image package.
